requirements.txt
dash
plotly
pandas
psycopg2-binary
SQLAlchemy
python-dotenv